requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
googlemaps>=4.5.1
pandas>=1.3.0
